feat: rm @ponder/core and using newest package ponder v0.11.2, intro multichain config
