+++ package.json
@@ -4,7 +4,7 @@
 	"private": false,
 	"license": "MIT",
 	"scripts": {
-		"dev": "ponder dev",
+		"dev": "yarn install && ponder dev",
 		"start": "ponder start",
 		"codegen": "ponder codegen",
 		"lint": "eslint .",
@@ -13,7 +13,7 @@
 	},
 	"dependencies": {
 		"@frankencoin/zchf": "^0.2.17",
-		"@ponder/core": "^0.7.17",
+		"ponder": "^0.11.2",
 		"hono": "^4.6.14",
 		"viem": "^2.13.8"
 	},
