+++ src/api/index.ts
@@ -1,0 +1,5 @@
+import { Hono } from 'hono';
+
+const app = new Hono();
+
+export default app;
